Checking in a bunch of test code to make sure SVN doesn't get too out of data. The modifications address:
1. Edited the contents of the scripts folder to automate switching back and forth between debug/retail functionality depending on which shortcut the person uses to launch the program. (footnotereap.lnk uses log level 9 and *dbg.lnk uses 0)

2. Major edits to footnote.au3 include:

A. GetClip() now:
i. optionally strips carriage returns from clip item
ii. properly saves the clipboard at the start of the program and restores it at the end in CleanupExit() (no longer saves and replaces clipboard after each modification)
iii. squelched tempclip debug output for logger levels > $EUNHANDLED

B. GetCurrentURL() now uses {ALTDOWN}{ALTUP} style syntax over "!d"

C. GetDirectoryNameFromURL() fixed a bug where it would incorrectly return partial matches from longer ids (e.g. id 457 could be matched against 45712). This had the affect where it would return the wrong information. Also fixed a bug where it wouldn't replace & with and and double quotes with two single quotes.

D. CreateNewDirectory() now optionally takes a parameter for the name of the new directory. If no parameter it uses GetDirectoryNameFromURL()

E. Added IsSameAsClip() used by SetSaveDialogDirectory(). This checks the output from the textbox to make sure it matches what was pasted in (an attempt to address to issue 6). There are numerous other bugs in this section related to issue 14 and issue 18.

F. StartDownloadImage() had a major revamp. Since StartDownloadImage has to be smart and figure out whether a person is continuing from where they previously left off or if they moved to a new URL. So at the start of the program it has to GetURLFromDirectory() and try to make sense of whether it matches the last registry entry. If it gets numerous matches where it's ambiguous it lets the user choose. If there's only one match, but the directory exists or doesn't exist it figures out the right thing to do. If there are no previous registry entries or no directories that match it just creates a new one. I also found a case where occasionally it doesn't navigate to the correct directory. This could cause an infinite loop. So to get around this after it fails to navigate to the new folder after two tries it navigates up one and then back down to the original folder. 

G. StartResume() now gets an @ERROR message to let it know it started at a random url that wasn't page 1. 

H. For my own sanity I now add a build number to each iteration. So when I compare logs I know whether I'm comparing apples to apples or apples to oranges. Anyone who submits a bug should always be using build #1. Typically I'll be the only one with a later build.

I. Abstracted numerous magic numbers and literal strings to constants or variables that can be either set at the beginning of the program or in the registry.